move wait for port function to test-utils.ts
