+++ packages/protocol/lib/test-utils.ts
@@ -2,7 +2,7 @@ import { hasEntryInRegistry, usesRegistry } from '@celo/protocol/lib/registry-ut
 import BigNumber from 'bignumber.js'
 import * as chai from 'chai'
 import * as chaiSubset from 'chai-subset'
-import { spawn } from 'child_process'
+import { spawn, SpawnOptions } from 'child_process'
 import { keccak256 } from 'ethereumjs-util'
 import {
   ProxyInstance,
@@ -107,6 +107,41 @@ export async function exec(command: string, args: string[]) {
       }
     })
   })
+}
+
+export function execCmd(
+  cmd: string,
+  args: string[],
+  options?: SpawnOptions & { silent?: boolean }
+) {
+  return new Promise<number>(async (resolve, reject) => {
+    const { silent, ...spawnOptions } = options || { silent: false }
+    if (!silent) {
+      console.debug('$ ' + [cmd].concat(args).join(' '))
+    }
+    const process = spawn(cmd, args, { ...spawnOptions, stdio: silent ? 'ignore' : 'inherit' })
+    process.on('close', (code) => {
+      try {
+        resolve(code)
+      } catch (error) {
+        reject(error)
+      }
+    })
+  })
+}
+
+async function isPortOpen(host: string, port: number) {
+  return (await execCmd('nc', ['-z', host, port.toString()], { silent: true })) === 0
+}
+
+async function waitForPortOpen(host: string, port: number, seconds: number) {
+  const deadline = Date.now() + seconds * 1000
+  do {
+    if (await isPortOpen(host, port)) {
+      return true
+    }
+  } while (Date.now() < deadline)
+  return false
 }
 
 export const assertProxiesSet = async (getContract: any) => {
